auto slot fixes
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@341 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ windows/Wizard.h
@@ -48,8 +48,8 @@ public:
 
 	void write();
 	void setLang();
-	void setDownloadLimits(int value);
-	void setUploadLimits(int value);
+	void setDownloadLimits(double value);
+	void setUploadLimits(double value);
 
 private:
 	CComboBox ctrlUpload;
